# swagger-trailing-slash.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Paperless Rest Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/documents/": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "GetDocuments",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "query",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ordering",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags__id__all",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "document_type__id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "storage_path__id__in",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "correspondent__id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "truncate_content",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
				  "type": "object",
				  "properties": {
					"count": {
					  "type": "integer"
					},
					"next": {
					  "type": "integer"
					},
					"previous": {
					  "type": "integer"
					},
					"all": {
					  "type": "array",
					  "items": 
						{
						  "type": "integer"
						}
					},
					"results": {
					  "type": "array",
					  "items": 
						{
						  "type": "object",
						  "properties": {
							"id": {
							  "type": "integer"
							},
							"correspondent": {
							  "type": "integer"
							},
							"document_type": {
							  "type": "integer"
							},
							"storage_path": {
							  "type": "integer"
							},
							"title": {
							  "type": "string"
							},
							"content": {
							  "type": "string"
							},
							"tags": {
							  "type": "array",
							  "items": 
								{
								  "type": "integer"
								}
							},
							"created": {
							  "type": "string"
							},
							"created_date": {
							  "type": "string"
							},
							"modified": {
							  "type": "string"
							},
							"added": {
							  "type": "string"
							},
							"archive_serial_number": {
							  "type": "integer"
							},
							"original_file_name": {
							  "type": "string"
							},
							"archived_file_name": {
							  "type": "string"
							},
							"owner": {
							  "type": "integer"
							},
							"user_can_change": {
							  "type": "boolean"
							},
							"notes": {
							  "type": "array",
							  "items": 
								{
								  "type": "object",
								  "properties": {
									"id": {
									  "type": "integer"
									},
									"note": {
									  "type": "string"
									},
									"created": {
									  "type": "string"
									},
									"document": {
									  "type": "integer"
									},
									"user": {
									  "type": "integer"
									}
								  },
								  "required": [
									"id",
									"note",
									"created",
									"document",
									"user"
								  ]
								}
							}
						  },
						  "required": [
							"id",
							"correspondent",
							"document_type",
							"storage_path",
							"title",
							"content",
							"tags",
							"created",
							"created_date",
							"modified",
							"added",
							"archive_serial_number",
							"original_file_name",
							"archived_file_name",
							"owner",
							"user_can_change",
							"notes"
						  ]
						}
					}
				  },
				  "required": [
					"count",
					"next",
					"previous",
					"all",
					"results"
				  ]
				},
                "examples": {
                  "default": {
                    "value": "{\"count\":1,\"next\":null,\"previous\":null,\"all\":[1],\"results\":[{\"id\":1,\"correspondent\":1,\"document_type\":1,\"storage_path\":1,\"title\":\"test\",\"content\":\"test\",\"tags\":[1],\"created\":\"2023-09-08T00:00:00Z\",\"created_date\":\"2023-09-08\",\"modified\":\"2023-09-08T12:15:24.652098Z\",\"added\":\"2023-09-08T12:05:28.338799Z\",\"archive_serial_number\":1,\"original_file_name\":\"test.pdf\",\"archived_file_name\":\"2023-09-08 edtest test.pdf\",\"owner\":3,\"user_can_change\":true,\"notes\":[{\"id\":1,\"note\":\"pdf in testdir\",\"created\":\"2023-09-08T12:11:05.848131Z\",\"document\":1,\"user\":3},{\"id\":2,\"note\":\"pdf 2 in test\",\"created\":\"2023-09-08T12:15:12.328015Z\",\"document\":1,\"user\":3}]}]}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "GetDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "full_perms",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
				  "type": "object",			
				  "properties": {
					"id": {
					  "type": "integer"
					},
					"correspondent": {
					  "type": "integer"
					},
					"document_type": {
					  "type": "integer"
					},
					"storage_path": {
					  "type": "integer"
					},
					"title": {
					  "type": "string"
					},
					"content": {
					  "type": "string"
					},
					"tags": {
					  "type": "array",
					  "items": 
						{
						  "type": "integer"
						}
					},
					"created": {
					  "type": "string"
					},
					"created_date": {
					  "type": "string"
					},
					"modified": {
					  "type": "string"
					},
					"added": {
					  "type": "string"
					},
					"archive_serial_number": {
					  "type": "integer"
					},
					"original_file_name": {
					  "type": "string"
					},
					"archived_file_name": {
					  "type": "string"
					},
					"owner": {
					  "type": "integer"
					},
					"permissions": {
					  "type": "object",
					  "properties": {
						"view": {
						  "type": "object",
						  "properties": {
							"users": {
							  "type": "array",
							  "items": 
								{
								  "type": "integer"
								}
							},
							"groups": {
							  "type": "array",
							  "items": 
								{
								  "type": "integer"
								}
							}
						  },
						  "required": [
							"users",
							"groups"
						  ]
						},
						"change": {
						  "type": "object",
						  "properties": {
							"users": {
							  "type": "array",
							  "items": 
								{
								  "type": "integer"
								}
							},
							"groups": {
							  "type": "array",
							  "items": 
								{
								  "type": "integer"
								}
							}
						  },
						  "required": [
							"users",
							"groups"
						  ]
						}
					  },
					  "required": [
						"view",
						"change"
					  ]
					},
					"notes": {
					  "type": "array",
					  "items": 
						{
						  "type": "object",
						  "properties": {
							"id": {
							  "type": "integer"
							},
							"note": {
							  "type": "string"
							},
							"created": {
							  "type": "string"
							},
							"document": {
							  "type": "integer"
							},
							"user": {
							  "type": "integer"
							}
						  },
						  "required": [
							"id",
							"note",
							"created",
							"document",
							"user"
						  ]
						}
					}
				  },
				  "required": [
					"id",
					"correspondent",
					"document_type",
					"storage_path",
					"title",
					"content",
					"tags",
					"created",
					"created_date",
					"modified",
					"added",
					"archive_serial_number",
					"original_file_name",
					"archived_file_name",
					"owner",
					"permissions",
					"notes"
				  ]
				},
                "examples": {
                  "default": {
                    "value": "{\"id\":1,\"correspondent\":1,\"document_type\":1,\"storage_path\":1,\"title\":\"test\",\"content\":\"test\",\"tags\":[1],\"created\":\"2023-09-08T00:00:00Z\",\"created_date\":\"2023-09-08\",\"modified\":\"2023-09-08T12:15:24.652098Z\",\"added\":\"2023-09-08T12:05:28.338799Z\",\"archive_serial_number\":1,\"original_file_name\":\"test.pdf\",\"archived_file_name\":\"2023-09-08 edtest test.pdf\",\"owner\":3,\"permissions\":{\"view\":{\"users\":[3],\"groups\":[1]},\"change\":{\"users\":[3],\"groups\":[1]}},\"notes\":[{\"id\":1,\"note\":\"pdf in testdir\",\"created\":\"2023-09-08T12:11:05.848131Z\",\"document\":1,\"user\":3},{\"id\":2,\"note\":\"pdf 2 in test\",\"created\":\"2023-09-08T12:15:12.328015Z\",\"document\":1,\"user\":3}]}"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Documents"
        ],
        "operationId": "UpdateDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                  "type": "object",
				  "properties": {
					"id": {
					  "type": "integer"
					},
					"correspondent": {
					  "type": "integer"
					},
					"document_type": {
					  "type": "integer"
					},
					"storage_path": {
					  "type": "integer"
					},
					"title": {
					  "type": "string"
					},
					"content": {
					  "type": "string"
					},
					"tags": {
					  "type": "array",
					  "items": 
						{
						  "type": "integer"
						}
					},
					"created_date": {
					  "type": "string"
					},
					"modified": {
					  "type": "string"
					},
					"added": {
					  "type": "string"
					},
					"archive_serial_number": {
					  "type": "integer"
					},
					"original_file_name": {
					  "type": "string"
					},
					"archived_file_name": {
					  "type": "string"
					},
					"owner": {
					  "type": "integer"
					},
					"permissions": {
					  "type": "object",
					  "properties": {
						"view": {
						  "type": "object",
						  "properties": {
							"users": {
							  "type": "array",
							  "items": {}
							},
							"groups": {
							  "type": "array",
							  "items": {}
							}
						  },
						  "required": [
							"users",
							"groups"
						  ]
						},
						"change": {
						  "type": "object",
						  "properties": {
							"users": {
							  "type": "array",
							  "items": {}
							},
							"groups": {
							  "type": "array",
							  "items": {}
							}
						  },
						  "required": [
							"users",
							"groups"
						  ]
						}
					  },
					  "required": [
						"view",
						"change"
					  ]
					},
					"notes": {
					  "type": "array",
					  "items": {}
					},
					"set_permissions": {
					  "type": "object",
					  "properties": {
						"view": {
						  "type": "object",
						  "properties": {
							"users": {
							  "type": "array",
							  "items": {}
							},
							"groups": {
							  "type": "array",
							  "items": {}
							}
						  },
						  "required": [
							"users",
							"groups"
						  ]
						},
						"change": {
						  "type": "object",
						  "properties": {
							"users": {
							  "type": "array",
							  "items": {}
							},
							"groups": {
							  "type": "array",
							  "items": {}
							}
						  },
						  "required": [
							"users",
							"groups"
						  ]
						}
					  },
					  "required": [
						"view",
						"change"
					  ]
					}
				  },
				  "required": [
					"id",
					"correspondent",
					"document_type",
					"storage_path",
					"title",
					"content",
					"tags",
					"created_date",
					"modified",
					"added",
					"archive_serial_number",
					"original_file_name",
					"archived_file_name",
					"owner",
					"permissions",
					"notes",
					"set_permissions"
				  ]
              },
                "examples": {
                  "default": {
                    "value": "{\"id\":2,\"correspondent\":null,\"document_type\":null,\"storage_path\":null,\"title\":\"toom BauMarkt GmbH Kassenzettel 15.07.2023 (1)\",\"content\":\"toomft toomit\nRespekt, wet’s selber macht.\ntoqm Baumarkt GmbH Respekt, wer’s selber macht\nGotti leb-Daimler-Str. 6\n28816 Stuhr-Brinkum\n0421 258430 toom Baumarkt GmbH\nNutzen Sie schon unsere Gott 1 i eb-Dai fitj’er-Str. 6\nVorteilskarte? 28816 Stuhr-Brinkum\n0421 258430\n4388858925116 EINKAUFSKAR-TQN . Nutzen Sie schon unsere\n1.000 STK\n0.99 19 Vorteilskarte?\n4004014794012 Innenspachtel 1\n1.000 STK 4388858925116 EINKAUFSKART0N\n3.89 19 1.000 STK\n4004624146300 i-lT Plexschlaueh 0.99 19\n1.000 STK 9.99 19 4004014794012 Innenspachtel 1\n1.000 STK\n4006209942598 Uni versa Idiibel ni 3.89 19\n2 000 STK a 3.99 7.98 19 4004624146300 HT Flexschiauch\n1.000 STK\n4013307727064 Nahtroller 9.99 19\n1.000 STK 4.99 19 4006209942598 Uni versaidubel ra\n4048962238969 DU0P0WER 2.000 STK a 3.99 7.98 19\n1.000 STK 4013307727064 Nahtroller\n3. 89 19 1.000 STK\n4048962238952 DU0P0KER 4. 89 19\n1.000 STK 4048962238969 EMPOWER\n3.99 19 1.000 STK\n4032526917230 FI achverbi iides' 3.99 19\n3.000 STK a 2.09 6.27 19 4048962238952 DU0PQWER\n1.000 STK\n4337256145251 SPS Pan. 5. fix 20 3 99 1.9\n1.000 STK 4032526917230 Flachverbinder\n6.19 IS\n4032526756006 K-Scheme DIN 90 3.000 STK a 2.09 6.27 19\n1.000 STK 4337256145251 SPS Pan. 5, Ox 20\n6.19 19 1.000 STK\n4032526657556 Aufh-Ose 24x 36 6.19 19\n1.000 STK 4032526756006 K-Schetbe DIN 90\n3.19 19 1.000 STK\n4002593412020 Gipserspachtei 6.19 19\n1.000 STK 4032526657566 Aufh-Ose 24x 36\n8.29 19 1.000 STK\n4006596162186 Gabelstuck verch 3.19 19\n1.000 STK 4002593412020 Gipserspachte1\n7.99 19 1.000 STK\n4006596161206 Doppein1ppe1 8.29 19\n2.000 STK a 3.59 7.18 19 4006596162166 Gabel stuck verch\n40G6596255042 Gewinded1 chiband 1.000 STK\n1.000 STK 7.99 19\n3.29 19 4006596161206 Doppe1nippe!\n4004624846200 Getaieanschlusss 2.000 STK a 3.59 7.18 19\n1.000 STK 8.98 19 4006596255042 Gewindedichiband\n4007591143020 Schneiispachtel 1.000 STK\n1.000 STK 3.29 19\n4.99 19 4004624846200 Get ateansch1usss\n4004014781166 Premium Buntlack 1.000 STK\n1.000 STK 8.99 19\n4.99 19 4007591143020 Schneiispachtel\n4013307727231 Abbrechklingen 1.000 STK\n1.000 STK 4.99 19\n6.29 19 4004014781166 Premium Bunt lack\n4015000428217 Pattex Montage P 1.000 STK\n1.000 STK 4.99 19\n7.99 19 4013307727231 Abbrechklingen\n00000071101OS API 410x60x3, Bern 1.000 STK\n6.28 19\n4.100 LFH a 29.99 122.96 19 4015000428211 Pattex Montage P\n1.000 STK\nSUMME 7.99 19\nEUR 240.62 0000007110109 APL 410x60x3, Scat\n4. TOO LFM a 29.99 122.96 19\nGEGEBEN EC-Cash 240.62\nSUMME\nEUR 240.62\nDatum:\n* * Kundenbeleg * *\n15.07.2023 GEGEBEN EC-Cash\nUhrzeit: 240. 62\n10:47:17 Uhr\nBeleg-Nr.\nTrace-Nr.\n6751\nDatum-' * * Kundenbeleg * *\n726487 15. 07. 2023\n—Kar terizahl Uhrzeit:\nBeleg-Nr. 10:47:17 Uhr\nContactless 6751\ngirocard Trace-Nr,\nNr. 726487\nTerminal-ID *#«MMiw«MM8O00 oooo Kartenzahlung\n56014757 Contactless\nPos-Info girocard\n00 075 00\nAS-Zeit 15 07. 10:47 Uhr\nNr.\nBet rag EUR Termina1 -ID WW«W##i?ft80QQ 0000\n240,62 56014757\nZahlung erfolgt Pos-Info\nAS-Zeit 15.07. 00 075 00\nMINST CODE Bet rag EUR 10:47 Uhr\n19.00 % NEITO-Entgelt HWST-Betray 240,62\n202.20 38. 42 Zahlung erfolgt\nSteuer.Nr. : 215/5940/2539 MUST CODE\nNETTO-Entye It\nUID Nr. : DE812720447 19.00 %\n202.20\n\nSteuer.Nr.:\nUID 215/5940/2539\nNr ' nroio-wnii-\",\"tags\":[1],\"created_date\":\"2023-07-15\",\"modified\":\"2023-09-08T18:23:17.751946Z\",\"added\":\"2023-09-08T18:23:17.470212Z\",\"archive_serial_number\":34,\"original_file_name\":\"toom BauMarkt GmbH Kassenzettel 15.07.2023 (1).pdf\",\"archived_file_name\":\"2023-07-15 toom BauMarkt GmbH Kassenzettel 15.07.2023 (1).pdf\",\"owner\":3,\"permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}},\"notes\":[],\"set_permissions\":{\"view\":{\"users\":[],\"groups\":[]},\"change\":{\"users\":[],\"groups\":[]}}}"
                  }
                }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
				  "type": "object",
				  "properties": {
					"id": {
					  "type": "integer"
					},
					"correspondent": {
					  "type": "integer"
					},
					"document_type": {
					  "type": "integer"
					},
					"storage_path": {
					  "type": "integer"
					},
					"title": {
					  "type": "string"
					},
					"content": {
					  "type": "string"
					},
					"tags": {
					  "type": "array",
					  "items": 
						{
						  "type": "integer"
						}
					},
					"created": {
					  "type": "string"
					},
					"created_date": {
					  "type": "string"
					},
					"modified": {
					  "type": "string"
					},
					"added": {
					  "type": "string"
					},
					"archive_serial_number": {
					  "type": "integer"
					},
					"original_file_name": {
					  "type": "string"
					},
					"archived_file_name": {
					  "type": "string"
					},
					"owner": {
					  "type": "integer"
					},
					"user_can_change": {
					  "type": "boolean"
					},
					"notes": {
					  "type": "array",
					  "items": {}
					}
				  },
				  "required": [
					"id",
					"correspondent",
					"document_type",
					"storage_path",
					"title",
					"content",
					"tags",
					"created",
					"created_date",
					"modified",
					"added",
					"archive_serial_number",
					"original_file_name",
					"archived_file_name",
					"owner",
					"user_can_change",
					"notes"
				  ]
				},
                "examples": {
                  "default": {
                    "value": "{\"id\":2,\"correspondent\":null,\"document_type\":null,\"storage_path\":null,\"title\":\"toom BauMarkt GmbH Kassenzettel 15.07.2023 (1)\",\"content\":\"toomft toomit\nRespekt, wet’s selber macht.\ntoqm Baumarkt GmbH Respekt, wer’s selber macht\nGotti leb-Daimler-Str. 6\n28816 Stuhr-Brinkum\n0421 258430 toom Baumarkt GmbH\nNutzen Sie schon unsere Gott 1 i eb-Dai fitj’er-Str. 6\nVorteilskarte? 28816 Stuhr-Brinkum\n0421 258430\n4388858925116 EINKAUFSKAR-TQN . Nutzen Sie schon unsere\n1.000 STK\n0.99 19 Vorteilskarte?\n4004014794012 Innenspachtel 1\n1.000 STK 4388858925116 EINKAUFSKART0N\n3.89 19 1.000 STK\n4004624146300 i-lT Plexschlaueh 0.99 19\n1.000 STK 9.99 19 4004014794012 Innenspachtel 1\n1.000 STK\n4006209942598 Uni versa Idiibel ni 3.89 19\n2 000 STK a 3.99 7.98 19 4004624146300 HT Flexschiauch\n1.000 STK\n4013307727064 Nahtroller 9.99 19\n1.000 STK 4.99 19 4006209942598 Uni versaidubel ra\n4048962238969 DU0P0WER 2.000 STK a 3.99 7.98 19\n1.000 STK 4013307727064 Nahtroller\n3. 89 19 1.000 STK\n4048962238952 DU0P0KER 4. 89 19\n1.000 STK 4048962238969 EMPOWER\n3.99 19 1.000 STK\n4032526917230 FI achverbi iides' 3.99 19\n3.000 STK a 2.09 6.27 19 4048962238952 DU0PQWER\n1.000 STK\n4337256145251 SPS Pan. 5. fix 20 3 99 1.9\n1.000 STK 4032526917230 Flachverbinder\n6.19 IS\n4032526756006 K-Scheme DIN 90 3.000 STK a 2.09 6.27 19\n1.000 STK 4337256145251 SPS Pan. 5, Ox 20\n6.19 19 1.000 STK\n4032526657556 Aufh-Ose 24x 36 6.19 19\n1.000 STK 4032526756006 K-Schetbe DIN 90\n3.19 19 1.000 STK\n4002593412020 Gipserspachtei 6.19 19\n1.000 STK 4032526657566 Aufh-Ose 24x 36\n8.29 19 1.000 STK\n4006596162186 Gabelstuck verch 3.19 19\n1.000 STK 4002593412020 Gipserspachte1\n7.99 19 1.000 STK\n4006596161206 Doppein1ppe1 8.29 19\n2.000 STK a 3.59 7.18 19 4006596162166 Gabel stuck verch\n40G6596255042 Gewinded1 chiband 1.000 STK\n1.000 STK 7.99 19\n3.29 19 4006596161206 Doppe1nippe!\n4004624846200 Getaieanschlusss 2.000 STK a 3.59 7.18 19\n1.000 STK 8.98 19 4006596255042 Gewindedichiband\n4007591143020 Schneiispachtel 1.000 STK\n1.000 STK 3.29 19\n4.99 19 4004624846200 Get ateansch1usss\n4004014781166 Premium Buntlack 1.000 STK\n1.000 STK 8.99 19\n4.99 19 4007591143020 Schneiispachtel\n4013307727231 Abbrechklingen 1.000 STK\n1.000 STK 4.99 19\n6.29 19 4004014781166 Premium Bunt lack\n4015000428217 Pattex Montage P 1.000 STK\n1.000 STK 4.99 19\n7.99 19 4013307727231 Abbrechklingen\n00000071101OS API 410x60x3, Bern 1.000 STK\n6.28 19\n4.100 LFH a 29.99 122.96 19 4015000428211 Pattex Montage P\n1.000 STK\nSUMME 7.99 19\nEUR 240.62 0000007110109 APL 410x60x3, Scat\n4. TOO LFM a 29.99 122.96 19\nGEGEBEN EC-Cash 240.62\nSUMME\nEUR 240.62\nDatum:\n* * Kundenbeleg * *\n15.07.2023 GEGEBEN EC-Cash\nUhrzeit: 240. 62\n10:47:17 Uhr\nBeleg-Nr.\nTrace-Nr.\n6751\nDatum-' * * Kundenbeleg * *\n726487 15. 07. 2023\n—Kar terizahl Uhrzeit:\nBeleg-Nr. 10:47:17 Uhr\nContactless 6751\ngirocard Trace-Nr,\nNr. 726487\nTerminal-ID *#«MMiw«MM8O00 oooo Kartenzahlung\n56014757 Contactless\nPos-Info girocard\n00 075 00\nAS-Zeit 15 07. 10:47 Uhr\nNr.\nBet rag EUR Termina1 -ID WW«W##i?ft80QQ 0000\n240,62 56014757\nZahlung erfolgt Pos-Info\nAS-Zeit 15.07. 00 075 00\nMINST CODE Bet rag EUR 10:47 Uhr\n19.00 % NEITO-Entgelt HWST-Betray 240,62\n202.20 38. 42 Zahlung erfolgt\nSteuer.Nr. : 215/5940/2539 MUST CODE\nNETTO-Entye It\nUID Nr. : DE812720447 19.00 %\n202.20\n\nSteuer.Nr.:\nUID 215/5940/2539\nNr ' nroio-wnii-\"\",\"tags\":[1],\"created\":\"2023-07-15T00:00:00Z\",\"created_date\":\"2023-07-15\",\"modified\":\"2023-09-08T19:40:43.833922Z\",\"added\":\"2023-09-08T18:23:17.470212Z\",\"archive_serial_number\":34,\"original_file_name\":\"toom BauMarkt GmbH Kassenzettel 15.07.2023 (1).pdf\",\"archived_file_name\":\"2023-07-15 toom BauMarkt GmbH Kassenzettel 15.07.2023 (1).pdf\",\"owner\":3,\"user_can_change\":true,\"notes\":[]}"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "operationId": "DeleteDocument",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/documents/post_document/": {
      "post": {
        "tags": [
          "Documents"
        ],
        "operationId": "UploadDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "created": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "document_type": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int32"
                    }
                  },
                  "correspondent": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "document": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              },
              "encoding": {
                "title": {
                  "style": "form"
                },
                "created": {
                  "style": "form"
                },
                "document_type": {
                  "style": "form"
                },
                "tags": {
                  "style": "form"
                },
                "correspondent": {
                  "style": "form"
                },
                "document": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}